chore: [IA-91] Contextual help event (#3329)
* Extract ContextualInfo component

* first implementation?

* Refactor navigation utils to leverage current types

* Use FC and add some tests

* Remove unused imports and fix utils

* Make naming of ContextualInfo consistent

* Apply minor review suggestions

* Ensure edge cases don't break navigation helper

* Track event only for non-blacklisted routes

* refactoring
